+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Added support for `solium-disable-previous-line` comment directive.
 - Added support for `solium-enable` comment directive. See [configuring with comments](https://ethlint.readthedocs.io/en/latest/user-guide.html#configuring-with-comments). This feature currently has a limitation which has been documented in [Known Issues](https://ethlint.readthedocs.io/en/latest/known-issues.html).
 - Fixed rule `no-empty-blocks` to report function declarations with empty bodies. Fallback and `payable` functions and `payable` constructors are not reported if their body is empty. See [#254](https://github.com/duaraghav8/Ethlint/issues/254).
+- Fixed rule `quotes` to stop reporting false positives due to brackets enclosing strings (see [#240](https://github.com/duaraghav8/Ethlint/issues/240)).
 - Modified rule `uppercase` to allow up to 2 leading and trailing underscores for a constant's name.
 
 ## 1.2.2 (2019-01-13)